Rename Function Files and Add Comments. Remove unused code
